+++ app/Http/Controllers/Frontend/TaskCard/TaskCardRoutineController.php
@@ -232,8 +232,8 @@ class TaskCardRoutineController extends Controller
             $taskCard->accesses()->sync($request->access);
             $taskCard->zones()->sync($request->zone);
             $taskCard->related_to()->sync($request->relationship);
-            $taskCard->thresholds()->delete();
-            $taskCard->repeats()->delete();
+            if($taskCard->thresholds)$taskCard->thresholds()->delete();
+            if($taskCard->repeats) $taskCard->repeats()->delete();
             for ($i=0; $i < sizeof($request->threshold_amount) ; $i++) { 
                 $taskCard->thresholds()->save(new Threshold([
                     'type_id' => Type::where('uuid',$request->threshold_type[$i])->first()->id,
